Add repository interfaces and refactor
