+++ customer/src/main/java/nl/softwarestrijders/waiter/customer/core/domain/exceptions/InvalidAdditionException.java
@@ -1,9 +1,9 @@
-package nl.softwarestrijders.waiter.customer.domain.exceptions;
+package nl.softwarestrijders.waiter.customer.core.domain.exceptions;
 
 /**
  * Exception that is thrown when the given addition doesn't comply with the set boundaries for additions
  *
- * @see nl.softwarestrijders.waiter.customer.domain.Address#setAddition(String) addition
+ * @see nl.softwarestrijders.waiter.customer.core.domain.Address#setAddition(String) addition
  */
 public class InvalidAdditionException extends RuntimeException {
 	public InvalidAdditionException() {
