+++ customer/src/main/java/nl/softwarestrijders/waiter/customer/core/domain/exceptions/InvalidCityException.java
@@ -1,9 +1,9 @@
-package nl.softwarestrijders.waiter.customer.domain.exceptions;
+package nl.softwarestrijders.waiter.customer.core.domain.exceptions;
 
 /**
  * Exception that is thrown when the given city doesn't comply with the set boundaries for city naming
  *
- * @see nl.softwarestrijders.waiter.customer.domain.Address#setCity(String) city
+ * @see nl.softwarestrijders.waiter.customer.core.domain.Address#setCity(String) city
  */
 public class InvalidCityException extends RuntimeException {
 	public InvalidCityException() {
