+++ customer/src/main/java/nl/softwarestrijders/waiter/customer/core/domain/exceptions/InvalidEmailException.java
@@ -1,9 +1,9 @@
-package nl.softwarestrijders.waiter.customer.domain.exceptions;
+package nl.softwarestrijders.waiter.customer.core.domain.exceptions;
 
 /**
  * Exception that is thrown if the email address contains invalid characters or if the composition contains failures.
  *
- * @see nl.softwarestrijders.waiter.customer.domain.Customer#setEmail (String) (String) email address
+ * @see nl.softwarestrijders.waiter.customer.core.domain.Customer#setEmail (String) (String) email address
  */
 public class InvalidEmailException extends RuntimeException {
 	public InvalidEmailException() {
