+++ customer/src/main/java/nl/softwarestrijders/waiter/customer/core/domain/exceptions/InvalidHouseNumberException.java
@@ -1,9 +1,9 @@
-package nl.softwarestrijders.waiter.customer.domain.exceptions;
+package nl.softwarestrijders.waiter.customer.core.domain.exceptions;
 
 /**
  * Exception that is thrown when the given house number doens't comply with the set house number boundaries
  *
- * @see nl.softwarestrijders.waiter.customer.domain.Address#setHouseNumber(int) houseNumber
+ * @see nl.softwarestrijders.waiter.customer.core.domain.Address#setHouseNumber(int) houseNumber
  */
 public class InvalidHouseNumberException extends RuntimeException {
 	public InvalidHouseNumberException() {
