+++ customer/src/main/java/nl/softwarestrijders/waiter/customer/core/domain/exceptions/InvalidNameException.java
@@ -1,10 +1,10 @@
-package nl.softwarestrijders.waiter.customer.domain.exceptions;
+package nl.softwarestrijders.waiter.customer.core.domain.exceptions;
 
 
 /**
  * Exception that is thrown if the name contains invalid characters
  *
- * @see nl.softwarestrijders.waiter.customer.domain.Customer customer
+ * @see nl.softwarestrijders.waiter.customer.core.domain.Customer customer
  */
 public class InvalidNameException extends RuntimeException {
 	public InvalidNameException(char c) {
