+++ customer/src/main/java/nl/softwarestrijders/waiter/customer/core/domain/exceptions/InvalidPostalCodeException.java
@@ -1,9 +1,9 @@
-package nl.softwarestrijders.waiter.customer.domain.exceptions;
+package nl.softwarestrijders.waiter.customer.core.domain.exceptions;
 
 /**
  * Exception that is thrown when the given postal code doesn't comply with the set boundaries for postal codes
  *
- * @see nl.softwarestrijders.waiter.customer.domain.Address#setPostalCode(String) postalCode
+ * @see nl.softwarestrijders.waiter.customer.core.domain.Address#setPostalCode(String) postalCode
  */
 public class InvalidPostalCodeException extends RuntimeException {
 	public InvalidPostalCodeException(String cause) {
