+++ customer/src/main/java/nl/softwarestrijders/waiter/customer/storage/CustomerRepository.java
@@ -1,0 +1,7 @@
+package nl.softwarestrijders.waiter.customer.storage;
+
+import nl.softwarestrijders.waiter.customer.core.domain.Customer;
+import org.springframework.data.mongodb.repository.MongoRepository;
+
+public interface CustomerRepository extends MongoRepository<Customer, String> {
+}
